R3 AR-1: extract native-SQL emission into a submodule
Move the parser_override native-SQL rewrite cluster out of parse/mod.rs
into parse::native_sql:

- rewrite_to_native_sql — the dispatch entry (CREATE/DROP/ALTER ->
  native INSERT/DELETE/UPDATE; DESCRIBE/SHOW -> passthrough).
- emit_native_create_sql / emit_native_create_from_yaml_file — the two
  CREATE emission paths (inline definition and FROM YAML FILE).
- rewrite_drop / rewrite_alter_rename / rewrite_alter_comment — the
  DROP/ALTER DML emitters with their pure-SQL existence/collision guards.
- escape_sql_arg / unescape_sql_arg and the three *_guard_select builders
  — compiled unconditionally so cargo test exercises escaping/guard
  wording without the extension stubs.

The FFI entry points (sv_parser_override_rust, sv_parse_function_rust,
sv_make/drop_override_context, run_validation_for_parse_function) stay in
mod.rs; they reach rewrite_to_native_sql via a `#[cfg(feature =
"extension")]` re-export. The escape/guard helpers are re-exported under
`#[cfg(test)]` for the parent module's unit tests. Import gating mirrors
the function gating so both bundled and extension builds stay
warning-clean.

No behaviour change. Verified: just build (extension), cargo test (1045),
clippy -D warnings, and just test-sql (64/64) all pass.
